+++ src/main/java/com/braintreegateway/util/Http.java
@@ -248,13 +248,17 @@ public class Http {
         writer.flush();
 
         FileInputStream inputStream = new FileInputStream(uploadFile);
-        byte[] buffer = new byte[4096];
-        int bytesRead;
-        while ((bytesRead = inputStream.read(buffer)) != -1) {
-            outputStream.write(buffer, 0, bytesRead);
+        try {
+            byte[] buffer = new byte[4096];
+            int bytesRead;
+            while ((bytesRead = inputStream.read(buffer)) != -1) {
+                outputStream.write(buffer, 0, bytesRead);
+            }
+            outputStream.flush();
         }
-        outputStream.flush();
-        inputStream.close();
+        finally {
+            inputStream.close();
+        }
 
         writer.append(LINE_FEED);
         writer.flush();