+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## unreleased
 * Add `lastFour` to `PaymentMethodNonceDetails`
+* Remove final specification from `MerchantAccountGateway` (#64)
 
 ## 2.83.1
 * Restore Javadoc