+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## unreleased
+## 2.84.0
 * Add `lastFour` to `PaymentMethodNonceDetails`
 * Remove final specification from `MerchantAccountGateway` (#64)
 * Fix dispute results in transactions not showing the correct status sometimes